# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HOP 주변인 코칭 API",
    "version": "0.1.0",
    "description": "Node.js 24 및 PostgreSQL 기반 백엔드입니다. 팀 초대 가입과 내부 시연용 임시 방문자 진입을 지원합니다. 정보 확보 수치는 인터뷰 진행 상태이며 진단·임상 척도가 아닙니다."
  },
  "servers": [
    {
      "url": "/",
      "description": "OpenAPI 문서를 제공하는 동일 서버입니다. 운영 도메인과 HTTPS 상태는 별도로 확인합니다."
    },
    {
      "url": "http://127.0.0.1:9000",
      "description": "로컬 개발 기본 주소"
    }
  ],
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/demo": {
      "get": {
        "operationId": "getDemoAvailability",
        "tags": ["인증"],
        "summary": "가입 없는 내부 시연 가능 여부",
        "security": [],
        "responses": {"200": {"description": "enabled는 HOP_DEMO_ENABLED와 v2 API 사용 가능 여부입니다."}}
      },
      "post": {
        "operationId": "startDemoVisit",
        "tags": ["인증"],
        "summary": "24시간 임시 방문자 대화 시작",
        "description": "HOP_DEMO_ENABLED=true인 내부 시연에서만 사용합니다. 방문자별 임시 계정과 동의 기록, 프로젝트, 첫 대화를 만들고 HttpOnly 세션 쿠키를 설정합니다. 빈 JSON 객체를 전송합니다. 세션은 24시간 뒤 만료되고 계정은 시간당 정리 작업에서 삭제됩니다.",
        "security": [],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "maxProperties": 0}}}},
        "responses": {"201": {"description": "user, project_id, conversation_id 및 세션 쿠키"}, "404": {"description": "시연 모드 비활성화"}, "429": {"$ref": "#/components/responses/RateLimited"}}
      },
      "delete": {
        "operationId": "deleteDemoVisit",
        "tags": ["인증"],
        "summary": "현재 임시 방문자와 온라인 기록 즉시 삭제",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "maxProperties": 0}}}},
        "responses": {"200": {"description": "deleted=true 및 세션 쿠키 삭제"}, "401": {"$ref": "#/components/responses/Unauthorized"}, "403": {"description": "시연 계정이 아님"}}
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "registerTeamMember",
        "tags": [
          "인증"
        ],
        "summary": "초대된 팀원 가입",
        "description": "HOP_ALLOW_REGISTRATION=true이고 유효한 팀 초대 코드가 있을 때만 가입합니다. 공개 가입은 지원하지 않습니다. 비밀번호는 신규 가입 시 12~128자입니다.",
        "security": [],
        "requestBody": {
          "required": true,
          "description": "64 KiB 이하 JSON 본문",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "가입 및 로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "HttpOnly; Path=/ 세션 쿠키. SameSite는 HOP_COOKIE_SAME_SITE(strict 기본값)이며 None은 Secure가 필수입니다. 로그인·변경은 7일, 로그아웃은 Max-Age=0.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "서버 요청 추적 ID. 대화 request_id와 별개입니다."
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "인증"
        ],
        "summary": "이메일·비밀번호 로그인",
        "description": "로그인 시 현재 요청에 기존 세션이 있으면 폐기하고 새 세션을 발급합니다.",
        "security": [],
        "requestBody": {
          "required": true,
          "description": "64 KiB 이하 JSON 본문",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "HttpOnly; Path=/ 세션 쿠키. SameSite는 HOP_COOKIE_SAME_SITE(strict 기본값)이며 None은 Secure가 필수입니다. 로그인·변경은 7일, 로그아웃은 Max-Age=0.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "서버 요청 추적 ID. 대화 request_id와 별개입니다."
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "tags": [
          "인증"
        ],
        "summary": "현재 사용자",
        "responses": {
          "200": {
            "description": "로그인한 사용자",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "서버 요청 추적 ID. 대화 request_id와 별개입니다."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "tags": [
          "인증"
        ],
        "summary": "현재 세션 폐기",
        "requestBody": {
          "required": true,
          "description": "빈 JSON 객체를 전송합니다.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "현재 세션 폐기됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logged_out": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "logged_out"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "HttpOnly; Path=/ 세션 쿠키. SameSite는 HOP_COOKIE_SAME_SITE(strict 기본값)이며 None은 Secure가 필수입니다. 로그인·변경은 7일, 로그아웃은 Max-Age=0.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "서버 요청 추적 ID. 대화 request_id와 별개입니다."
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "$ref": "#/components/responses/PayloadTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "description": " 쿠키 인증은 허용된 Origin이 필요합니다. Bearer 인증은 Origin을 생략할 수 있습니다."
      }
    },
    "/api/questions": {
      "get": {
        "operationId": "listQuestions",
        "tags": [
          "인터뷰"
        ],
        "summary": "주변인 인터뷰 17문항",
        "description": "현재 서비스의 인터뷰 문항이며 검증된 증상 점수 척도가 아닙니다.",
        "responses": {
          "200": {
            "description": "문항과 scored_scale:false",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Questions"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "서버 요청 추적 ID. 대화 request_id와 별개입니다."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/projects/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "get": {
        "operationId": "getProject",
        "tags": [
          "프로젝트"
        ],
        "summary": "프로젝트와 정보 확보 상태",
        "description": "다른 사용자의 프로젝트는 반환하지 않습니다.",
        "responses": {
          "200": {
            "description": "프로젝트",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "서버 요청 추적 ID. 대화 request_id와 별개입니다."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "delete": {
        "operationId": "deleteProject",
        "tags": [
          "프로젝트"
        ],
        "summary": "프로젝트와 종속 데이터 삭제",
        "description": "프로젝트의 대화, 프로필 버전, 요청 캐시도 PostgreSQL의 종속 관계에 따라 삭제됩니다. 쿠키 인증은 허용된 Origin이 필요합니다. Bearer 인증은 Origin을 생략할 수 있습니다.",
        "responses": {
          "200": {
            "description": "삭제됨",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deleted": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "deleted"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "서버 요청 추적 ID. 대화 request_id와 별개입니다."
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/InvalidInput"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/openapi.json": {
      "get": {
        "operationId": "getOpenApiDocument",
        "tags": [
          "상태"
        ],
        "summary": "OpenAPI 3.1 JSON 조회",
        "security": [],
        "responses": {
          "200": {
            "description": "이 API의 OpenAPI 명세",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            },
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "description": "서버 요청 추적 ID. 대화 request_id와 별개입니다."
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hop_session",
        "description": "브라우저 기본 인증. HttpOnly, SameSite=Strict, 7일 만료. HTTPS에서는 Secure. UI는 localStorage 토큰을 사용하지 않습니다."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "명령줄 등 비브라우저 클라이언트용 opaque 세션 토큰. JWT가 아닙니다."
      }
    },
...
    "responses": {
      "BadRequest": {
        "description": "잘못된 JSON 문법",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "유효한 로그인 세션이 없거나 로그인 정보가 일치하지 않음",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "RateLimited": {
        "description": "현재 프로세스의 요청 제한을 초과함",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "const": 60
            },
            "description": "재시도 전 대기 시간(초)"
          }
        }
      },
...
    }
  }
}